# Faster kill_garbage