--- a/ethcore/account-state/src/state.rs
+++ b/ethcore/account-state/src/state.rs
@@ -760,19 +760,21 @@ impl<B: Backend> State<B> {
 
 	/// Remove any touched empty or dust accounts.
 	pub fn kill_garbage(&mut self, touched: &HashSet<Address>, min_balance: &Option<U256>, kill_contracts: bool) -> TrieResult<()> {
-		touched.iter().for_each(|address| { // Check all touched accounts
-			if let Some(entry) = self.cache.borrow().get(address) {
-				if entry.exists_and_is_null() // Remove all empty touched accounts.
-					|| min_balance.map_or(false, |ref balance| entry.account.as_ref().map_or(false, |account|
-						(account.is_basic() || kill_contracts) // Remove all basic and optionally contract accounts where balance has been decreased.
-						&& account.balance() < balance && entry.old_balance.as_ref().map_or(false, |b| account.balance() < b))) {
-					// todo[dvdplm] This will call self.cache.borrow_mut() – is that ok here?
-					//  Yes(?), because self.cache is `RefCell<T>` so this is a single threaded call
-					//  and nobody else can call `borrow()` here?
-					self.insert_cache(address, AccountEntry::new_dirty(None));
-				}
-			}
-		});
+		let to_kill: HashSet<_> =
+			touched.iter().filter_map(|address| { // Check all touched accounts
+				if let Some(entry) = self.cache.borrow().get(address) {
+					if entry.exists_and_is_null() // Remove all empty touched accounts.
+						|| min_balance.map_or(false, |ref balance| entry.account.as_ref().map_or(false, |account|
+							(account.is_basic() || kill_contracts) // Remove all basic and optionally contract accounts where balance has been decreased.
+							&& account.balance() < balance && entry.old_balance.as_ref().map_or(false, |b| account.balance() < b))) {
+						Some(address)
+					} else { None }
+				} else { None }
+			}).collect();
+
+		for address in to_kill {
+			self.kill_account(address)
+		}
 		Ok(())
 	}