Profiling the [benchmark from #11509](https://github.com/paritytech/parity-ethereum/pull/11509) shows that almost half the time `apply()`-ing a transaction is spent in garbage collection. This PR avoids visiting each cache item and `collect()`-ing accounts to clean up.

```
apply transactions with tracing (Constantinople)
                        time:   [771.87 us 785.31 us 799.39 us]
                        change: [-1.6171% +0.8871% +3.4408%] (p = 0.49 > 0.05)
                        No change in performance detected.
Found 6 outliers among 100 measurements (6.00%)
  6 (6.00%) high mild

apply transactions without tracing (Constantinople)
                        time:   [624.40 us 632.38 us 640.60 us]
                        change: [-3.7894% -1.7770% +0.2331%] (p = 0.09 > 0.05)
                        No change in performance detected.

apply transactions with tracing (Istanbul)
                        time:   [603.13 us 611.19 us 619.69 us]
                        change: [+0.6069% +3.1495% +5.9820%] (p = 0.02 < 0.05)
                        Change within noise threshold.
Found 1 outliers among 100 measurements (1.00%)
  1 (1.00%) high mild

apply transactions without tracing (Istanbul)
                        time:   [490.90 us 498.22 us 505.93 us]
                        change: [-0.6020% +1.4944% +3.7714%] (p = 0.19 > 0.05)
                        No change in performance detected.
Found 3 outliers among 100 measurements (3.00%)
  3 (3.00%) high mild
 ```
 
 Before this PR:
 
 ```
 apply transactions with tracing (Constantinople)
                         time:   [2.0166 ms 2.0470 ms 2.0807 ms]
                         change: [-1.1628% +1.1355% +3.8393%] (p = 0.38 > 0.05)
                         No change in performance detected.
 Found 12 outliers among 100 measurements (12.00%)
   12 (12.00%) high mild

 Benchmarking apply transactions without tracing (Constantinople): Warming up for 3.0000 s
 Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 9.7s or reduce sample count to 50.
 apply transactions without tracing (Constantinople)
                         time:   [1.8491 ms 1.8758 ms 1.9056 ms]
                         change: [-3.5325% -1.0900% +1.4900%] (p = 0.40 > 0.05)
                         No change in performance detected.
 Found 10 outliers among 100 measurements (10.00%)
   6 (6.00%) high mild
   4 (4.00%) high severe

 Benchmarking apply transactions with tracing (Istanbul): Warming up for 3.0000 s
 Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 6.4s or reduce sample count to 60.
 apply transactions with tracing (Istanbul)
                         time:   [1.2220 ms 1.2451 ms 1.2708 ms]
                         change: [-6.6992% -3.7312% -0.7610%] (p = 0.02 < 0.05)
                         Change within noise threshold.
 Found 8 outliers among 100 measurements (8.00%)
   6 (6.00%) high mild
   2 (2.00%) high severe

 Benchmarking apply transactions without tracing (Istanbul): Warming up for 3.0000 s
 Warning: Unable to complete 100 samples in 5.0s. You may wish to increase target time to 5.7s or reduce sample count to 60.
 apply transactions without tracing (Istanbul)
                         time:   [1.0961 ms 1.1153 ms 1.1382 ms]
                         change: [-1.6498% +0.7117% +3.0664%] (p = 0.57 > 0.05)
                         No change in performance detected.
 Found 14 outliers among 100 measurements (14.00%)
   6 (6.00%) high mild
   8 (8.00%) high severe
 ```