@ordian 2020-02-24
>  No change in performance detected.

But if there is no perf improvement, what's the point?