@ordian 2020-02-24
> > No change in performance detected.
> 
> But if there is no perf improvement, what's the point?

Ah, sorry, now I see it's actually faster, should have looked at the absolute numbers.

Test failure is legit though.